modularize schmackbone
