+++ .eslintrc
@@ -1,98 +1,177 @@
 {
   "parserOptions": {
-    "ecmaVersion": 8,
-    "ecmaFeatures": {
-      "experimentalObjectRestSpread": true,
-      "impliedStrict": true
-    }
+    "ecmaVersion": 2018,
+    "ecmaFeatures": {"impliedStrict": true}
   },
   "env": {
     "browser": true,
-    "node": true,
+    "commonjs": true,
     "es6": true,
-    "amd": true
-  },
-  "globals": {
-    "attachEvent": false,
-    "detachEvent": false
+    "node": true
   },
   "rules": {
-    "array-bracket-spacing": 2,
-    "block-scoped-var": 2,
-    "brace-style": [1, "1tbs", {"allowSingleLine": true}],
-    "camelcase": 2,
+    "array-bracket-newline": [2, "consistent"],
+    "array-bracket-spacing": [2, "never"],
+    "array-callback-return": 2,
+    "arrow-body-style": [2, "as-needed"],
+    "arrow-parens": 2,
+    "arrow-spacing": 2,
+    "brace-style": [2, "1tbs"],
     "comma-dangle": [2, "never"],
-    "comma-spacing": 2,
+    "comma-spacing": [2, {"after": true}],
+    "comma-style": [2, "last"],
     "computed-property-spacing": [2, "never"],
-    "dot-notation": [2, { "allowKeywords": true }],
+    "consistent-this": [2, "mountedInstance"],
+    "curly": 2,
+    "default-case": 2,
+    "dot-location": [2, "property"],
+    "dot-notation": 2,
     "eol-last": 2,
-    "eqeqeq": [2, "smart"],
-    "indent": [2, 2, {"SwitchCase": 1, "VariableDeclarator": 2}],
-    "key-spacing": 1,
-    "keyword-spacing": [2, { "after": true }],
+    "eqeqeq": 2,
+    "func-call-spacing": 2,
+    "generator-star-spacing": ["error", {"before": false, "after": true}],
+    "guard-for-in": 2,
+    "id-length": [2, {"exceptions": ["i", "x", "y", "_", "$"], "properties": "never"}],
+    "indent": [2, 2, {
+      "flatTernaryExpressions": true,
+      "MemberExpression": 2,
+      "SwitchCase": 1,
+      "VariableDeclarator": {"var": 2, "let": 2, "const": 3}
+    }],
+    "jsx-quotes": [2, "prefer-single"],
+    "key-spacing": [2, {"beforeColon": false, "afterColon": true}],
+    "keyword-spacing": [2, {}],
+    "line-comment-position": [2, {"position": "above"}],
     "linebreak-style": 2,
-    "max-depth": [1, 4],
-    "max-params": [1, 5],
-    "new-cap": [2, {"newIsCapExceptions": ["model"]}],
-    "no-alert": 2,
-    "no-caller": 2,
-    "no-catch-shadow": 2,
+    "lines-between-class-members": 2,
+    "max-len": [2, 100],
+    "max-statements-per-line": 2,
+    "new-cap": 2,
+    "no-bitwise": 2,
+    "no-case-declarations": 2,
+    "no-class-assign": 2,
+    "no-cond-assign": [2, "except-parens"],
     "no-console": 2,
+    "no-const-assign": 2,
     "no-debugger": 2,
     "no-delete-var": 2,
-    "no-div-regex": 1,
-    "no-dupe-args": 2,
+    "no-dupe-class-members": 2,
     "no-dupe-keys": 2,
     "no-duplicate-case": 2,
-    "no-else-return": 1,
-    "no-empty-character-class": 2,
+    "no-duplicate-imports": 2,
+    "no-else-return": 2,
+    "no-empty": 2,
+    "no-empty-pattern": 2,
     "no-eval": 2,
-    "no-ex-assign": 2,
     "no-extend-native": 2,
     "no-extra-boolean-cast": 2,
-    "no-extra-parens": 1,
     "no-extra-semi": 2,
     "no-fallthrough": 2,
     "no-floating-decimal": 2,
     "no-func-assign": 2,
+    "no-global-assign": 2,
     "no-implied-eval": 2,
     "no-inner-declarations": 2,
-    "no-irregular-whitespace": 2,
-    "no-label-var": 2,
-    "no-labels": 2,
-    "no-lone-blocks": 2,
     "no-lonely-if": 2,
-    "no-multi-str": 2,
-    "no-native-reassign": 2,
-    "no-negated-in-lhs": 1,
+    "no-loop-func": 2,
+    "no-mixed-spaces-and-tabs": 2,
+    "no-multi-spaces": 2,
+    "no-multiple-empty-lines": [2, {"max": 1, "maxEOF": 0, "maxBOF": 0}],
     "no-new-object": 2,
+    "no-new-symbol": 2,
     "no-new-wrappers": 2,
-    "no-obj-calls": 2,
-    "no-octal": 2,
-    "no-octal-escape": 2,
     "no-proto": 2,
-    "no-redeclare": 2,
+    "no-restricted-imports": [2, {
+      "paths": ["lodash", "d3"]
+    }],
+    "no-return-await": 2,
     "no-shadow": 2,
     "no-spaced-func": 2,
+    "no-tabs": 2,
+    "no-template-curly-in-string": 2,
+    "no-this-before-super": 2,
     "no-throw-literal": 2,
     "no-trailing-spaces": 2,
     "no-undef": 2,
     "no-undef-init": 2,
-    "no-undefined": 2,
-    "no-unneeded-ternary": 2,
+    "no-unneeded-ternary": [2, {"defaultAssignment": false}],
     "no-unreachable": 2,
-    "no-unused-expressions": [2, {"allowTernary": true, "allowShortCircuit": true}],
+    "no-unused-vars": [2, {"vars": "local", "args": "none"}],
+    "no-use-before-define": [2, "nofunc"],
+    "no-useless-computed-key": 2,
+    "no-useless-constructor": 2,
+    "no-useless-escape": 2,
+    "no-useless-rename": 2,
+    "no-useless-return": 2,
+    "no-void": 2,
     "no-with": 2,
-    "object-curly-spacing": [2, "never"],
-    "quote-props": [1, "consistent-as-needed"],
-    "quotes": [2, "single", "avoid-escape"],
+    "no-whitespace-before-property": 2,
+    "object-curly-newline": [2, {"multiline": true}],
+    "object-curly-spacing": 2,
+    "object-property-newline": [2, {"allowMultiplePropertiesPerLine": true}],
+    "object-shorthand": [2, "properties"],
+    "one-var-declaration-per-line": [2, "always"],
+    "operator-linebreak": [2, "after", {"overrides": {":": "ignore"}}],
+    "padded-blocks": [2, "never"],
+    "padding-line-between-statements": [2,
+      {blankLine: "always", prev: ["const", "let", "var"], next: "*"},
+      {blankLine: "any", prev: ["const", "let", "var"], next: ["const", "let", "var"]},
+      {blankLine: "always", prev: "*", next: "return"},
+      {blankLine: "always", prev: "*", next: "block-like"},
+      {blankLine: "always", prev: "block-like", next: "*"},
+      {blankLine: "never", prev: "case", next: "*"},
+      {blankLine: "never", prev: "*", next: "case"},
+      {blankLine: "always", prev: "function", next: "*"},
+      {blankLine: "always", prev: "import", next: "*"},
+      {blankLine: "any", prev: "import", next: "import"}
+    ],
+    "prefer-arrow-callback": 2,
+    "prefer-spread": 2,
+    "quote-props": [2, "as-needed"],
+    "quotes": [2, "single"],
     "radix": 2,
-    "semi": 2,
-    "space-before-function-paren": [2, {"anonymous": "never", "named": "never"}],
-    "space-infix-ops": 2,
-    "space-unary-ops": [2, { "words": true, "nonwords": false }],
+    "rest-spread-spacing": 2,
+    "semi": [2, "always"],
+    "semi-spacing": 2,
+    "sift/array-element-newline": [2, {"multiline": true}],
+    "sift/function-paren-newline": [2, "multiline"],
+    "sift/jsx-handler-names": 2,
+    "sift/space-infix-ops": 2,
+    "sort-imports": [2, {"ignoreCase": true}],
+    "space-before-blocks": [2, "always"],
+    "space-before-function-paren": [2, "never"],
+    "space-in-parens": 2,
+    "space-unary-ops": [2, {"words": true, "nonwords": false}],
+    "spaced-comment": [2, "always", {
+      "exceptions": ["*"],
+      "markers": ["*"]
+    }],
+    "strict": [2, "global"],
+    "switch-colon-spacing": 2,
+    "symbol-description": 2,
+    "template-curly-spacing": 2,
+    "template-tag-spacing": 2,
     "use-isnan": 2,
     "valid-typeof": 2,
-    "wrap-iife": [2, "inside"]
-  }
+    "vars-on-top": 2,
+    "wrap-iife": [2, "inside"],
+    "yoda": [2, "never", {"exceptRange": true}]
+  },
+  "globals": {
+    "babel": false,
+    "Bloodhound": false,
+    "CodeMirror": false,
+    "confirm": false,
+    "fetch": false,
+    "FontFaceObserver": false,
+    "mapboxgl": false,
+    "mixpanel": false,
+    "mostRecentAjaxRequest": false,
+    "mpPageName": false,
+    "olark": false
+  },
+  "plugins": [
+    "sift"
+  ],
+  "parser": "babel-eslint"
 }
